--- src/aspect-meta-model/characteristic/default-values-collection.ts
@@ -1,8 +1,0 @@
-import { DefaultEntityInstance } from "../default-entity-instance";
-import { Characteristic } from "./default-characteristic";
-
-export interface ValuesCollection extends Characteristic {
-    values: Array<DefaultEntityInstance | string | number>;
-
-    indexOf(value: string): number;
-}
